feat(viewer): Image viewer is back
