+++ projects/knora/viewer/src/lib/view/resource-view/resource-view.component.html
@@ -5,15 +5,21 @@
 
         <div class="resource">
             <!-- media: show file representation -->
-            <!-- <div *ngIf="HasFileRepresentation">
-                <div [ngSwitch]="resource.fileRepresentationsToDisplay[0].type" class="media"></div>
-            </div> -->
-            <!-- <div *ngIf="resource.fileRepresentationsToDisplay?.stillImage && resource.fileRepresentationsToDisplay?.stillImage.length > 0"
-                 class="media">
-                <kui-still-image #kuiStillImage class="osd-viewer"
-                                 [images]="resource.fileRepresentationsToDisplay.stillImage">
-                </kui-still-image>
-            </div> -->
+            <div *ngIf="fileRepresentation">
+                <div [ngSwitch]="fileRepresentation[0].type" class="media">
+
+                    <kui-still-image *ngSwitchCase="Constants.StillImageFileValue" #kuiStillImage class="osd-viewer"
+                                     [images]="fileRepresentation">
+                    </kui-still-image>
+
+                    <kui-message *ngSwitchDefault [medium]="true"
+                                 [message]="{status: 501, statusMsg: 'Not yet implemented', statusText: fileRepresentation[0].type + ' is not yet implemented'}">
+                    </kui-message>
+
+                </div>
+            </div>
+
+            <!--  -->
             <!--
                 <div [ngSwitch]="resource.fileRepresentationsToDisplay[0].type" class="media">
                     <div *ngSwitchCase="KnoraConstants.StillImageFileValue">
@@ -27,9 +33,7 @@
                         <kui-moving-image></kui-moving-image>
                     </div>
 
-                    <div *ngSwitchDefault>
-                        <p>This media type {{resource.fileRepresentationsToDisplay[0].type}} is not yet implemented</p>
-                    </div>
+
                 </div>
                 -->
 
@@ -38,23 +42,21 @@
                 <!-- first tab for the main resource e.g. book -->
                 <mat-tab [label]="resource.entityInfo?.classes[resource.type].label">
                     <!-- header with project / ontology / resource class info / ark url -->
+                    <kui-properties-toolbar *ngIf="toolbar" [projectiri]="resource.attachedToProject"
+                                            [allProps]="allProps" (toggleProps)="toggleProps($event)">
+                    </kui-properties-toolbar>
 
-                    <kui-properties-view [resource]="resource" [toolbar]="toolbar" [allProps]="allProps">
+                    <kui-properties-view [propArray]="propArray" [toolbar]="toolbar" [allProps]="allProps">
                     </kui-properties-view>
 
-
-                    <!-- TODO: [guiOrder]="guiOrder" /  [ontologyInfo]="resource.entityInfo" is missing? couldn't find yet -->
-                    <!-- <kui-properties-view [properties]="resource?.properties" [entityInfo]="resource?.entityInfo"
-                                             [incomingLinks]="resource?.incomingReferences">
-                        </kui-properties-view> -->
                 </mat-tab>
 
                 <!-- TODO: second tab for a "part-of"-resource e.g. book page -->
                 <!-- Couldn't find an incomingFileRepresentation in the new ReadResource from knora-api-js-lib -->
-                <!-- <mat-tab *ngIf="resource.incomingReferences.length > 0 && currentResource"
-                        [label]="sequence.ontologyInformation.getLabelForResourceClass(currentResource.type)">
+                <mat-tab *ngIf="resource.incomingReferences.length > 0 && currentResource"
+                         [label]="sequence.ontologyInformation.getLabelForResourceClass(currentResource.type)">
 
-                </mat-tab> -->
+                </mat-tab>
 
                 <!-- TODO: third tab for a "region"-resource e.g. region on book page -->
                 <!-- <mat-tag></mat-tag> -->
